#50 organise tests
